+++ packages/uno-serverless-aws/test/unit/services/config-test.ts
@@ -2,7 +2,7 @@ import { AWSError, Response, SSM } from "aws-sdk";
 import { PromiseResult } from "aws-sdk/lib/request";
 import { expect } from "chai";
 import { describe, it } from "mocha";
-import { ConfigService, HealthCheckStatus } from "uno-serverless";
+import { ConfigService, HealthCheckStatus, StandardErrorCodes } from "uno-serverless";
 import { SSMParameterStoreClient, SSMParameterStoreConfigService } from "../../../src/services/config";
 
 const path = "/uno-serverless/tests";
@@ -68,7 +68,7 @@ describe("SSMParameterStoreConfigService", () => {
       await config.get("foo");
       expect(false);
     } catch (error) {
-      expect(error.code).to.equal("configurationError");
+      expect(error.code).to.equal(StandardErrorCodes.ConfigurationError);
       expect(error.data.key).to.equal("foo");
       expect(error.data.provider).to.equal("SSMParameterStoreConfigService");
     }
@@ -77,7 +77,7 @@ describe("SSMParameterStoreConfigService", () => {
       await config.get("foo", true);
       expect(false);
     } catch (error) {
-      expect(error.code).to.equal("configurationError");
+      expect(error.code).to.equal(StandardErrorCodes.ConfigurationError);
       expect(error.data.key).to.equal("foo");
       expect(error.data.provider).to.equal("SSMParameterStoreConfigService");
     }